+++ tests/mock.py
@@ -6,6 +6,7 @@ import logging
 import pathlib
 import socket
 import traceback
+import typing  # noqa: F401
 import urllib.error
 import urllib.parse
 import urllib.request
@@ -31,7 +32,7 @@ class FixtureOpener(urllib.request.OpenerDirector):
             urllib.parse.urljoin(base_url, './w/api.php')
         )
         # ./w/api.php?action=query&prop=imageinfo|info&inprop=url&iiprop=url|size|mime&format=json&titles={}  # noqa: E501
-        self.records = []
+        self.records = []  # type: typing.List[typing.Tuple[str, str]]
         cls = type(self)
         self.logger = logging.getLogger(cls.__qualname__) \
                              .getChild(cls.__name__)