+++ docs/changes.rst
@@ -6,8 +6,13 @@ Version 0.5.0
 
 To be released.
 
-- Added :mod:`wikidata.cache` module and :class:`~wikidata.cache.CachePolicy`
-  interface in it.
+- Wikidata API calls over network became possible to be cached.
+
+  - :class:`~wikidata.client.Client` now has
+    :attr:`~wikidata.client.Client.cache_policy` attribute and constructor
+    option.  Nothing is cached by default.
+  - Added :mod:`wikidata.cache` module and :class:`~wikidata.cache.CachePolicy`
+    interface in it.
 
 
 Version 0.4.1